# pkgs/openapi/dist/openapi.json
{
 "openapi": "3.0.0",
 "info": {
  "title": "myPocket",
  "version": "1.0.0"
 },
 "paths": {
...
  "/profile/status/": {
   "get": {
    "tags": [
     "profile"
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "email"
        ],
        "properties": {
         "email": {
          "type": "string",
          "example": "joeDue@gmail.com"
         }
        }
       }
      }
     }
    },
    "summary": "authentication status",
    "description": "when provided with email, this will resolve to\nall available authentication methods for given email\n",
    "responses": {
     "200": {
      "description": "Ok",
      "content": {
       "application/json": {
        "x-success": true,
        "examples": {
         "singed up": {
          "description": "singed up with multiple authentication methods",
          "value": [
           "local",
           "google"
          ]
         },
         "not signed up": {
          "description": "email is not signed up",
          "value": []
         }
        },
        "schema": {
         "type": "array",
         "items": {
          "type": "string",
          "example": [
           "local",
           "google"
          ],
          "enum": [
           "local",
           "google"
          ]
         }
        }
       }
      }
     },
     "400": {
      "description": "some of the fields are missing\n`these fields are required`\n",
      "content": {
       "application/json": {
        "x-failed": true,
        "schema": {
         "type": "object",
         "properties": {
          "status": {
           "type": "number",
           "default": 400
          },
          "message": {
           "description": "in form of `these fields are required\\:email`",
           "type": "string"
          },
          "name": {
           "type": "string",
           "default": "SomeFieldsRequired"
          },
          "details": {
           "type": "object",
           "properties": {
            "errors": {
             "type": "object",
             "description": "key value pair where key is properties\nname and value is more explanation\n"
            }
           }
          }
         }
        }
       }
      }
     }
    }
   }
  },
  "/profile/": {
   "get": {
    "security": [
     {
      "BearerToken": []
     }
    ],
    "parameters": [
     {
      "name": "Authorization",
      "in": "header",
      "description": "can be obtained by `auth/*/*`, in form of eyxxx.xxx.xxx",
      "required": true,
      "schema": {
       "type": "string",
       "example": "Bearer xxxxxxx="
      }
     }
    ],
    "tags": [
     "profile"
    ],
    "summary": "get profile",
    "description": "retrieve profile info",
    "responses": {
     "200": {
      "description": "Ok",
      "content": {
       "application/json": {
        "x-success": true,
        "schema": {
         "allOf": [
          {
           "type": "object",
           "required": [
            "_id",
            "__v"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "626819066b8ac48b612dda69"
            },
            "__v": {
             "oneOf": [
              {
               "type": "string",
               "enum": [
                "offline"
               ]
              },
              {
               "type": "number",
               "enum": [
                0
               ]
              }
             ],
             "example": 0,
             "default": 0,
             "enum": [
              0,
              "offline"
             ]
            }
           }
          },
          {
           "type": "object",
           "required": [
            "createdAt",
            "updatedAt"
           ],
           "properties": {
            "createdAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            },
            "updatedAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          },
          {
           "type": "object",
           "required": [
            "displayName",
            "email",
            "providers",
            "token"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "63da2a0a643dd3aa49f5c6b1"
            },
            "__v": {
             "type": "number",
             "example": 0,
             "default": 0
            },
            "displayName": {
             "type": "string",
             "example": "joe Due"
            },
            "email": {
             "type": "string",
             "example": "joeDue@gmail.com"
            },
            "token": {
             "type": "string",
             "example": "eyJxxxx.xxxxx.xxxxxx"
            },
            "providers": {
             "type": "array",
             "example": [
              "local",
              "google"
             ],
             "items": {
              "type": "string",
              "enum": [
               "local",
               "google"
              ]
             }
            },
            "picture": {
             "type": "string",
             "example": "https ://example.com/path/to/img"
            }
           }
          }
         ]
        }
       }
      }
     },
     "401": {
      "description": "authentication via bearer token has failed",
      "content": {
       "application/json": {
        "x-failed": true,
        "schema": {
         "type": "object",
         "properties": {
          "status": {
           "type": "number",
           "default": 401
          },
          "message": {
           "description": "more details",
           "example": "authentication failed",
           "type": "string"
          },
          "name": {
           "type": "string",
           "default": "TokenFailed"
          },
          "details": {
           "type": "object",
           "required": [
            "type"
           ],
           "properties": {
            "type": {
             "type": "string",
             "enum": [
              "JsonWebTokenError",
              "TokenExpiredError",
              "NoTokenWasProvided",
              "UnspecifiedError"
             ],
             "example": "TokenExpiredError"
            },
            "date": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          }
         }
        }
       }
      }
     }
    }
   },
   "put": {
    "security": [
     {
      "BearerToken": []
     }
    ],
    "parameters": [
     {
      "name": "Authorization",
      "in": "header",
      "description": "can be obtained by `auth/*/*`, in form of eyxxx.xxx.xxx",
      "required": true,
      "schema": {
       "type": "string",
       "example": "Bearer xxxxxxx="
      }
     }
    ],
    "requestBody": {
     "description": "fields to update",
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "displayName": {
          "type": "string",
          "example": "Joe Doe"
         },
         "picture": {
          "type": "string",
          "example": "https ://example.com/path/to/img\""
         }
        }
       }
      }
     }
    },
    "tags": [
     "profile"
    ],
    "summary": "update profile",
    "description": "update profile information",
    "responses": {
     "200": {
      "description": "Ok",
      "content": {
       "application/json": {
        "x-success": true,
        "schema": {
         "allOf": [
          {
           "type": "object",
           "required": [
            "_id",
            "__v"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "626819066b8ac48b612dda69"
            },
            "__v": {
             "oneOf": [
              {
               "type": "string",
               "enum": [
                "offline"
               ]
              },
              {
               "type": "number",
               "enum": [
                0
               ]
              }
             ],
             "example": 0,
             "default": 0,
             "enum": [
              0,
              "offline"
             ]
            }
           }
          },
          {
           "type": "object",
           "required": [
            "createdAt",
            "updatedAt"
           ],
           "properties": {
            "createdAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            },
            "updatedAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          },
          {
           "type": "object",
           "required": [
            "displayName",
            "email",
            "providers",
            "token"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "63da2a0a643dd3aa49f5c6b1"
            },
            "__v": {
             "type": "number",
             "example": 0,
             "default": 0
            },
            "displayName": {
             "type": "string",
             "example": "joe Due"
            },
            "email": {
             "type": "string",
             "example": "joeDue@gmail.com"
            },
            "token": {
             "type": "string",
             "example": "eyJxxxx.xxxxx.xxxxxx"
            },
            "providers": {
             "type": "array",
             "example": [
              "local",
              "google"
             ],
             "items": {
              "type": "string",
              "enum": [
               "local",
               "google"
              ]
             }
            },
            "picture": {
             "type": "string",
             "example": "https ://example.com/path/to/img"
            }
           }
          }
         ]
        }
       }
      }
     },
     "401": {
      "description": "authentication via bearer token has failed",
      "content": {
       "application/json": {
        "x-failed": true,
        "schema": {
         "type": "object",
         "properties": {
          "status": {
           "type": "number",
           "default": 401
          },
          "message": {
           "description": "more details",
           "example": "authentication failed",
           "type": "string"
          },
          "name": {
           "type": "string",
           "default": "TokenFailed"
          },
          "details": {
           "type": "object",
           "required": [
            "type"
           ],
           "properties": {
            "type": {
             "type": "string",
             "enum": [
              "JsonWebTokenError",
              "TokenExpiredError",
              "NoTokenWasProvided",
              "UnspecifiedError"
             ],
             "example": "TokenExpiredError"
            },
            "date": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          }
         }
        }
       }
      }
     }
    }
   }
  },
  "/profile/password/": {
   "put": {
    "security": [
     {
      "BearerToken": []
     }
    ],
    "parameters": [
     {
      "name": "Authorization",
      "in": "header",
      "description": "can be obtained by `auth/*/*`, in form of eyxxx.xxx.xxx",
      "required": true,
      "schema": {
       "type": "string",
       "example": "Bearer xxxxxxx="
      }
     }
    ],
    "requestBody": {
     "description": "update the password",
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "newPassword"
        ],
        "properties": {
         "oldPassword": {
          "description": "may be required if `local` authenticate is provided",
          "type": "string",
          "example": 12345567
         },
         "newPassword": {
          "type": "string",
          "example": "12345567differentPassword"
         }
        }
       }
      }
     }
    },
    "tags": [
     "profile"
    ],
    "summary": "(re)set password",
    "description": "update users password",
    "responses": {
     "200": {
      "description": "Ok",
      "content": {
       "application/json": {
        "x-success": true,
        "schema": {
         "allOf": [
          {
           "type": "object",
           "required": [
            "_id",
            "__v"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "626819066b8ac48b612dda69"
            },
            "__v": {
             "oneOf": [
              {
               "type": "string",
               "enum": [
                "offline"
               ]
              },
              {
               "type": "number",
               "enum": [
                0
               ]
              }
             ],
             "example": 0,
             "default": 0,
             "enum": [
              0,
              "offline"
             ]
            }
           }
          },
          {
           "type": "object",
           "required": [
            "createdAt",
            "updatedAt"
           ],
           "properties": {
            "createdAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            },
            "updatedAt": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          },
          {
           "type": "object",
           "required": [
            "displayName",
            "email",
            "providers",
            "token"
           ],
           "properties": {
            "_id": {
             "type": "string",
             "example": "63da2a0a643dd3aa49f5c6b1"
            },
            "__v": {
             "type": "number",
             "example": 0,
             "default": 0
            },
            "displayName": {
             "type": "string",
             "example": "joe Due"
            },
            "email": {
             "type": "string",
             "example": "joeDue@gmail.com"
            },
            "token": {
             "type": "string",
             "example": "eyJxxxx.xxxxx.xxxxxx"
            },
            "providers": {
             "type": "array",
             "example": [
              "local",
              "google"
             ],
             "items": {
              "type": "string",
              "enum": [
               "local",
               "google"
              ]
             }
            },
            "picture": {
             "type": "string",
             "example": "https ://example.com/path/to/img"
            }
           }
          }
         ]
        }
       }
      }
     },
     "400": {
      "description": "some of the fields are missing\n`these fields are required`\n",
      "content": {
       "application/json": {
        "x-failed": true,
        "schema": {
         "type": "object",
         "properties": {
          "status": {
           "type": "number",
           "default": 400
          },
          "message": {
           "description": "in form of `these fields are required\\:email`",
           "type": "string"
          },
          "name": {
           "type": "string",
           "default": "SomeFieldsRequired"
          },
          "details": {
           "type": "object",
           "properties": {
            "errors": {
             "type": "object",
             "description": "key value pair where key is properties\nname and value is more explanation\n"
            }
           }
          }
         }
        }
       }
      }
     },
     "401": {
      "description": "authentication via bearer token has failed",
      "content": {
       "application/json": {
        "x-failed": true,
        "schema": {
         "type": "object",
         "properties": {
          "status": {
           "type": "number",
           "default": 401
          },
          "message": {
           "description": "more details",
           "example": "authentication failed",
           "type": "string"
          },
          "name": {
           "type": "string",
           "default": "TokenFailed"
          },
          "details": {
           "type": "object",
           "required": [
            "type"
           ],
           "properties": {
            "type": {
             "type": "string",
             "enum": [
              "JsonWebTokenError",
              "TokenExpiredError",
              "NoTokenWasProvided",
              "UnspecifiedError"
             ],
             "example": "TokenExpiredError"
            },
            "date": {
             "type": "string",
             "example": "2022-04-26T16:08:38.276Z"
            }
           }
          }
         }
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
...
  "securitySchemes": {
   "BasicToken": {
    "type": "http",
    "scheme": "basic",
    "description": "use email and password to gain\nuser's profile along with the\ntoken.\n```\ncurl 'http://exmaple:0000/api/path/to' \\\n  --H 'Authorization: Basic ZGtAZy5jOjEyMw==' \\\n```\n"
   },
   "BearerToken": {
    "type": "http",
    "scheme": "bearer",
    "description": "this is the standard way to authenticate\nrequests, such a token can be obtained via\nemailAuth or googleAuth\n```\ncurl 'http://exmaple:0000/api/path/to' \\\n  --H 'Authorization: Bearer eyxxx.xxx.xxx' \\\n```\n"
   }
  },
...
 },
 "servers": [
  {
   "url": "https://localhost:{port}/api/{version}",
   "description": "development localhost servers",
   "variables": {
    "version": {
     "default": "v1"
    },
    "port": {
     "default": "8811"
    }
   }
  }
 ],
...
}